Some bugs in the proccess of fixing

+++ itemRecieved.php
@@ -1,13 +1,20 @@
 <?php
   include_once "db_connection.php";
+  session_start();
 
-  $itemID = $_GET['itemID'];
-  $matchedID = $_GET['matchedID'];
+  if(isset($_GET['itemID']) && isset($_GET['matchedID']))
+  {
 
-  $sqlItem = "DELETE FROM Items WHERE ID = $itemID";
-  $conn->query($sqlItem);
-  $sqlMatched = "DELETE FROM Matched WHERE ID = $matchedID";
-  $conn->query($sqlMatched);
-  header("Location: account.php");
+    $itemID = $_GET['itemID'];
+    $matchedID = $_GET['matchedID'];
+
+    $checkQuerry = "SELECT "
+
+    $sqlItem = "DELETE FROM Items WHERE ID = $itemID";
+    $conn->query($sqlItem);
+    $sqlMatched = "DELETE FROM Matched WHERE ID = $matchedID";
+    $conn->query($sqlMatched);
+    header("Location: account.php");
+  }
 ?>
